Correct issues with roundup function
Roundup is a templated function that takes two arguments of the same
type. There were several instances that were calling the function with
one argument that was signed and another that was unsigned. Some
compilers will error on this condition. Fix these instances by casting
to the correct type.

+++ include/mgard-x/DataRefactoring/DataRefactor.hpp
@@ -39,7 +39,8 @@ public:
     size_t workspace_size = 1;
     for (DIM d = 0; d < D; d++) {
       if (d == D - 1) {
-        workspace_size *= roundup((shape[d] + 2) * sizeof(T), pitch_size);
+        workspace_size *=
+            roundup((size_t)(shape[d] + 2) * sizeof(T), pitch_size);
       } else {
         workspace_size *= shape[d] + 2;
       }
